Renamed script files to match naming conventions of project, updated broken links and updated scripts for barrel export and index import.
