+++ src/js/api/auth/logout.js
