+++ src/js/listeners/auth/index.js
@@ -1,1 +1,3 @@
 export * from './setLoginFormListener.js';
+export * from './registerListener.js';
+export * from './logoutListener.js';
