+++ src/js/listeners/auth/logoutListener.js
@@ -1,6 +1,5 @@
-
-import { logout } from "../../api/auth/logoutSetupFunction.js"
+import { logout } from '../../api/auth/index.js';
 
 // The element id  is temporary. need to replace it with the right one.
 
-document.quaryselector("[data-auth=logout]").addEventListener("click", logout);
+document.quaryselector('[data-auth=logout]').addEventListener('click', logout);
